GitHub/requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
cachetools==4.0.0
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.1
Deprecated==1.2.9
google-api-core==1.16.0
google-api-python-client==1.8.0
google-auth==1.12.0
google-auth-httplib2==0.0.3
googleapis-common-protos==1.51.0
httplib2==0.17.0
idna==2.8
importlib-metadata==0.22
more-itertools==7.2.0
numpy==1.18.1
oauth2client==4.1.3
packaging==19.1
pandas==1.0.1
pluggy==0.13.0
protobuf==3.11.3
py==1.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyDrive==1.3.1
PyGithub==1.51
PyJWT==1.7.1
pyparsing==2.4.2
pytest==5.1.2
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
requests==2.22.0
rsa==4.0
selenium==3.141.0
six==1.12.0
uritemplate==3.0.1
urllib3==1.25.3
wcwidth==0.1.7
windows-curses==2.1.0
wrapt==1.12.1
zipp==0.6.0
